refactor: move unit test for payment method and capture method
